requirements.txt
pandas
psycopg2-binary
sqlalchemy
streamlit
plotly
python-dotenv